+++ src/Creacion/FactoriaArticulos.java
@@ -5,6 +5,7 @@
  */
 package Creacion;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 /**
  *
@@ -12,10 +13,10 @@ import java.math.BigInteger;
  */
 public interface FactoriaArticulos {
     
-    public void crearBebida(String etiqueta, BigInteger existencias, String nombre, BigInteger precio);
-    public void crearMojito(String etiqueta, String nombre, BigInteger precio);
-    public void crearMontado(String etiqueta, String nombre, BigInteger precio);
-    public void crearBocata(String etiqueta, String nombre, BigInteger precio);
-    public void crearOtro(String etiqueta, String nombre, BigInteger precio);
+    public void crearBebida(String etiqueta, BigInteger existencias, String nombre, BigDecimal precio);
+    public void crearMojito(String etiqueta, String nombre, BigDecimal precio);
+    public void crearMontado(String etiqueta, String nombre, BigDecimal precio);
+    public void crearBocata(String etiqueta, String nombre, BigDecimal precio);
+    public void crearOtro(String etiqueta, String nombre, BigDecimal precio);
     
 }
