+++ src/Creacion/FactoriaComida.java
@@ -2,6 +2,7 @@
 package Creacion;
 
 import AccesoModelo.FachadaArticulos;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 /**
@@ -13,21 +14,21 @@ public class FactoriaComida implements FactoriaArticulos {
     FachadaArticulos fachada = new FachadaArticulos();
     
     @Override
-    public void crearBebida(String etiqueta, BigInteger existencias, String nombre, BigInteger precio){
+    public void crearBebida(String etiqueta, BigInteger existencias, String nombre, BigDecimal precio){
     }
     @Override
-    public void crearMojito(String etiqueta, String nombre, BigInteger precio){
+    public void crearMojito(String etiqueta, String nombre, BigDecimal precio){
     }
     @Override
-    public void crearMontado(String etiqueta, String nombre, BigInteger precio){
+    public void crearMontado(String etiqueta, String nombre, BigDecimal precio){
         fachada.crearMontado(etiqueta, nombre, precio);
     }
     @Override
-    public void crearBocata(String etiqueta, String nombre, BigInteger precio){
+    public void crearBocata(String etiqueta, String nombre, BigDecimal precio){
         fachada.crearBocata(etiqueta, nombre, precio);
     }
     @Override
-    public void crearOtro(String etiqueta, String nombre, BigInteger precio){
+    public void crearOtro(String etiqueta, String nombre, BigDecimal precio){
         fachada.crearOtro(etiqueta, nombre, precio);
     }
 }
